move test to TestFunctionalIndex
remove the new class
