+++ hudi-spark-datasource/hudi-spark/src/test/scala/org/apache/spark/sql/hudi/command/index/TestFunctionalIndex.scala
@@ -19,32 +19,41 @@
 
 package org.apache.spark.sql.hudi.command.index
 
+import org.apache.hudi.DataSourceWriteOptions.{INSERT_OPERATION_OPT_VAL, OPERATION, PARTITIONPATH_FIELD, PRECOMBINE_FIELD, RECORDKEY_FIELD, TABLE_TYPE}
+import org.apache.hudi.HoodieConversionUtils.toProperties
 import org.apache.hudi.client.SparkRDDWriteClient
 import org.apache.hudi.client.common.HoodieSparkEngineContext
-import org.apache.hudi.{DataSourceReadOptions, HoodieFileIndex, HoodieSparkUtils}
-import org.apache.hudi.common.config.{HoodieMetadataConfig, TypedProperties}
+import org.apache.hudi.{DataSourceReadOptions, FunctionalIndexSupport, HoodieFileIndex, HoodieSparkUtils}
+import org.apache.hudi.common.config.{HoodieMetadataConfig, HoodieStorageConfig, TypedProperties}
+import org.apache.hudi.common.fs.FSUtils
 import org.apache.hudi.common.model.FileSlice
 import org.apache.hudi.common.table.HoodieTableMetaClient
+import org.apache.hudi.common.table.view.FileSystemViewManager
 import org.apache.hudi.common.testutils.HoodieTestUtils
 import org.apache.hudi.common.util.Option
-import org.apache.hudi.config.{HoodieCleanConfig, HoodieCompactionConfig, HoodieWriteConfig}
+import org.apache.hudi.config.{HoodieCleanConfig, HoodieCompactionConfig, HoodieIndexConfig, HoodieWriteConfig}
 import org.apache.hudi.hive.HiveSyncConfigHolder._
 import org.apache.hudi.hive.testutils.HiveTestUtil
 import org.apache.hudi.hive.{HiveSyncTool, HoodieHiveSyncClient}
+import org.apache.hudi.index.HoodieIndex
 import org.apache.hudi.metadata.{HoodieMetadataFileSystemView, MetadataPartitionType}
+import org.apache.hudi.storage.StoragePath
 import org.apache.hudi.sync.common.HoodieSyncConfig.{META_SYNC_BASE_PATH, META_SYNC_DATABASE_NAME, META_SYNC_NO_PARTITION_METADATA, META_SYNC_TABLE_NAME}
 import org.apache.hudi.testutils.HoodieClientTestUtils.createMetaClient
 import org.apache.hudi.util.JFunction
 import org.apache.spark.api.java.JavaSparkContext
+import org.apache.spark.sql.{Column, SaveMode}
 import org.apache.spark.sql.catalyst.analysis.{Analyzer, UnresolvedAttribute}
 import org.apache.spark.sql.catalyst.catalog.CatalogTable
-import org.apache.spark.sql.catalyst.expressions.{EqualTo, Expression, FromUnixTime, Literal}
+import org.apache.spark.sql.catalyst.expressions.{AttributeReference, EqualTo, Expression, FromUnixTime, Literal}
 import org.apache.spark.sql.catalyst.parser.ParserInterface
 import org.apache.spark.sql.hudi.command.{CreateIndexCommand, ShowIndexesCommand}
 import org.apache.spark.sql.hudi.common.HoodieSparkSqlTestBase
+import org.apache.spark.sql.types.{BinaryType, ByteType, DateType, DecimalType, IntegerType, ShortType, StringType, StructType, TimestampType}
 import org.junit.jupiter.api.Assertions.{assertEquals, assertTrue}
 import org.scalatest.Ignore
 
+import java.util.stream.Collectors
 import scala.collection.JavaConverters
 
 @Ignore
@@ -789,6 +798,163 @@ class TestFunctionalIndex extends HoodieSparkSqlTestBase {
         Seq("2021-09-26", "2021-09-26", false),
         Seq("2022-09-26", "2022-09-26", false)
       )
+    }
+  }
+
+  test("testFunctionalIndexUsingColumnStatsWithPartitionAndFilesFilter") {
+    if (HoodieSparkUtils.gteqSpark3_3) {
+      withTempDir { tmp =>
+        val tableName = generateTableName
+        val basePath = s"${tmp.getCanonicalPath}/$tableName"
+        val metadataOpts = Map(
+          HoodieMetadataConfig.ENABLE.key -> "true",
+          HoodieMetadataConfig.FUNCTIONAL_INDEX_ENABLE_PROP.key -> "true"
+        )
+        val opts = Map(
+          "hoodie.insert.shuffle.parallelism" -> "4",
+          "hoodie.upsert.shuffle.parallelism" -> "4",
+          HoodieWriteConfig.TBL_NAME.key -> tableName,
+          RECORDKEY_FIELD.key -> "c1",
+          PRECOMBINE_FIELD.key -> "c1",
+          PARTITIONPATH_FIELD.key() -> "c8"
+        )
+        val sourceJSONTablePath = getClass.getClassLoader.getResource("index/colstats/input-table-json-partition-pruning").toString
+
+        // NOTE: Schema here is provided for validation that the input date is in the appropriate format
+        val sourceTableSchema: StructType = new StructType()
+          .add("c1", IntegerType)
+          .add("c2", StringType)
+          .add("c3", DecimalType(9, 3))
+          .add("c4", TimestampType)
+          .add("c5", ShortType)
+          .add("c6", DateType)
+          .add("c7", BinaryType)
+          .add("c8", ByteType)
+        val inputDF = spark.read.schema(sourceTableSchema).json(sourceJSONTablePath)
+        inputDF
+          .sort("c1")
+          .repartition(4, new Column("c1"))
+          .write
+          .format("hudi")
+          .options(opts)
+          .option(HoodieStorageConfig.PARQUET_MAX_FILE_SIZE.key, 10 * 1024)
+          .option(OPERATION.key, INSERT_OPERATION_OPT_VAL)
+          .mode(SaveMode.Overwrite)
+          .save(basePath)
+        // Create a functional index on column c6
+        spark.sql(s"create table $tableName using hudi location '$basePath'")
+        spark.sql(s"create index idx_datestr on $tableName using column_stats(c6) options(func='identity')")
+        val metaClient = HoodieTableMetaClient.builder()
+          .setBasePath(basePath)
+          .setConf(HoodieTestUtils.getDefaultStorageConf)
+          .build()
+        assertTrue(metaClient.getTableConfig.getMetadataPartitions.contains("func_index_idx_datestr"))
+        assertTrue(metaClient.getIndexMetadata.isPresent)
+        assertEquals(1, metaClient.getIndexMetadata.get.getIndexDefinitions.size())
+
+        // check functional index records
+        val metadataConfig = HoodieMetadataConfig.newBuilder()
+          .fromProperties(toProperties(metadataOpts))
+          .build()
+        val functionalIndexSupport = new FunctionalIndexSupport(spark, metadataConfig, metaClient)
+        val prunedPartitions = Set("9")
+        var indexDf = functionalIndexSupport.loadFunctionalIndexDataFrame("func_index_idx_datestr", prunedPartitions, shouldReadInMemory = true)
+        // check only one record returned corresponding to the pruned partition
+        assertTrue(indexDf.count() == 1)
+        // select fileName from indexDf
+        val fileName = indexDf.select("fileName").collect().map(_.getString(0)).head
+        val fsv = FileSystemViewManager.createInMemoryFileSystemView(new HoodieSparkEngineContext(spark.sparkContext), metaClient,
+          HoodieMetadataConfig.newBuilder().enable(false).build())
+        fsv.loadAllPartitions()
+        val partitionPaths = fsv.getPartitionPaths
+        val partitionToBaseFiles: java.util.Map[String, java.util.List[StoragePath]] = new java.util.HashMap[String, java.util.List[StoragePath]]
+        // select file names for each partition from file system view
+        partitionPaths.forEach(partitionPath =>
+          partitionToBaseFiles.put(partitionPath.getName, fsv.getLatestBaseFiles(partitionPath.getName)
+            .map[StoragePath](baseFile => baseFile.getStoragePath).collect(Collectors.toList[StoragePath]))
+        )
+        fsv.close()
+        val expectedFileNames = partitionToBaseFiles.get(prunedPartitions.head).stream().map[String](baseFile => baseFile.getName).collect(Collectors.toSet[String])
+        assertTrue(expectedFileNames.size() == 1)
+        // verify the file names match
+        assertTrue(expectedFileNames.contains(fileName))
+
+        // check more records returned if no partition filter provided
+        indexDf = functionalIndexSupport.loadFunctionalIndexDataFrame("func_index_idx_datestr", Set(), shouldReadInMemory = true)
+        assertTrue(indexDf.count() > 1)
+      }
+    }
+  }
+
+  test("testComputeCandidateFileNames") {
+    if (HoodieSparkUtils.gteqSpark3_3) {
+      withTempDir { tmp =>
+        val tableName = generateTableName
+        val basePath = s"${tmp.getCanonicalPath}/$tableName"
+        // in this test, we will create a table with inserts going to log file so that there is a file slice with only log file and no base file
+        val metadataOpts = Map(
+          HoodieMetadataConfig.ENABLE.key -> "true",
+          HoodieMetadataConfig.FUNCTIONAL_INDEX_ENABLE_PROP.key -> "true"
+        )
+        val opts = Map(
+          "hoodie.insert.shuffle.parallelism" -> "4",
+          "hoodie.upsert.shuffle.parallelism" -> "4",
+          HoodieWriteConfig.TBL_NAME.key -> tableName,
+          TABLE_TYPE.key -> "MERGE_ON_READ",
+          RECORDKEY_FIELD.key -> "c1",
+          PRECOMBINE_FIELD.key -> "c1",
+          PARTITIONPATH_FIELD.key() -> "c8",
+          // setting IndexType to be INMEMORY to simulate Global Index nature
+          HoodieIndexConfig.INDEX_TYPE.key -> HoodieIndex.IndexType.INMEMORY.name()
+        )
+        val sourceJSONTablePath = getClass.getClassLoader.getResource("index/colstats/input-table-json-partition-pruning").toString
+
+        // NOTE: Schema here is provided for validation that the input date is in the appropriate format
+        val sourceTableSchema: StructType = new StructType()
+          .add("c1", IntegerType)
+          .add("c2", StringType)
+          .add("c3", DecimalType(9, 3))
+          .add("c4", TimestampType)
+          .add("c5", ShortType)
+          .add("c6", DateType)
+          .add("c7", BinaryType)
+          .add("c8", ByteType)
+        val inputDF = spark.read.schema(sourceTableSchema).json(sourceJSONTablePath)
+        inputDF
+          .sort("c1")
+          .repartition(4, new Column("c1"))
+          .write
+          .format("hudi")
+          .options(opts)
+          .option(HoodieStorageConfig.PARQUET_MAX_FILE_SIZE.key, 10 * 1024)
+          .option(OPERATION.key, INSERT_OPERATION_OPT_VAL)
+          .mode(SaveMode.Overwrite)
+          .save(basePath)
+        // Create a functional index on column c6
+        spark.sql(s"create table $tableName using hudi location '$basePath'")
+        spark.sql(s"create index idx_datestr on $tableName using column_stats(c6) options(func='identity')")
+        val metaClient = HoodieTableMetaClient.builder()
+          .setBasePath(basePath)
+          .setConf(HoodieTestUtils.getDefaultStorageConf)
+          .build()
+        assertTrue(metaClient.getTableConfig.getMetadataPartitions.contains("func_index_idx_datestr"))
+        assertTrue(metaClient.getIndexMetadata.isPresent)
+        assertEquals(1, metaClient.getIndexMetadata.get.getIndexDefinitions.size())
+        // check functional index records
+        val metadataConfig = HoodieMetadataConfig.newBuilder()
+          .fromProperties(toProperties(metadataOpts))
+          .build()
+        val fileIndex = new HoodieFileIndex(spark, metaClient, None,
+          opts ++ metadataOpts ++ Map("glob.paths" -> s"$basePath/9", DataSourceReadOptions.ENABLE_DATA_SKIPPING.key -> "true"), includeLogFiles = true)
+        val functionalIndexSupport = new FunctionalIndexSupport(spark, metadataConfig, metaClient)
+        val partitionFilter: Expression = EqualTo(AttributeReference("c8", IntegerType)(), Literal(9))
+        val (isPruned, prunedPaths) = fileIndex.prunePartitionsAndGetFileSlices(Seq.empty, Seq(partitionFilter))
+        assertTrue(isPruned)
+        val prunedPartitionAndFileNames = functionalIndexSupport.getPrunedPartitionsAndFileNames(prunedPaths, includeLogFiles = true)
+        assertTrue(prunedPartitionAndFileNames._1.size == 1) // partition
+        assertTrue(prunedPartitionAndFileNames._2.size == 1) // log file
+        assertTrue(FSUtils.isLogFile(prunedPartitionAndFileNames._2.head))
+      }
     }
   }
